+++ include/kernel/linux/unistd.h
@@ -35,7 +35,7 @@
 /* types */
 # ifndef off_t
 #  define off_t off_t
-typedef long long off_t;
+typedef long off_t;
 # endif
 # ifndef useconds_t
 #  define useconds_t useconds_t
